pyproject.toml
[project]
name = "saaf-compliance-shell"
version = "0.7.1"
description = "Modular compliance wrapper for AI agent workloads"
readme = "README.md"
license = "Apache-2.0"
license-files = ["LICENSE"]
requires-python = ">=3.11,<3.14"
dependencies = [
    "nemoguardrails>=0.10.0",
    "langchain-openai>=0.1.0",
    "presidio-analyzer>=2.2",
    "presidio-anonymizer>=2.2",
    "fastapi>=0.110",
    "uvicorn>=0.29",
    "httpx>=0.27",
    "openai>=1.30",
    "pyyaml>=6.0",
]

[project.scripts]
saaf-shell = "cli:main"

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "ruff>=0.6",
    "mypy>=1.10",
    "types-PyYAML>=6.0",
]
nlp = [
    "spacy>=3.7",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"

[tool.ruff]
line-length = 140
target-version = "py311"
extend-exclude = [".venv", "build", "dist"]

[tool.ruff.lint]
select = ["E", "F", "W", "I", "B", "UP", "SIM", "C4"]
# E501 is advisory — tests commonly hold long compliance fixtures.
# B008 would fire on FastAPI Depends(...) defaults; not relevant here yet.

[tool.ruff.lint.per-file-ignores]
"tests/*.py" = ["E501"]

[tool.mypy]
python_version = "3.11"
files = ["modules", "cli.py"]
ignore_missing_imports = true
follow_imports = "silent"
# Strictness is phased in — start with the cheap wins and tighten later.
warn_unused_ignores = true
warn_redundant_casts = true
no_implicit_optional = true
check_untyped_defs = true

[tool.setuptools]
py-modules = ["cli"]

[tool.setuptools.packages.find]
include = ["modules*", "guardrails*"]

[tool.setuptools.package-data]
guardrails = ["*.co", "*.yml"]

[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"
